Create README.md
X AutoCare is your absolute best source for used auto parts in the US. We believe in providing not only the best auto parts choices available to our customers but to provide excellent knowledgeable support and friendly service at every step from search to purchase. We aim to build solid relationships with our customers that go far and above our competitors to earn your trust and loyalty.

We are the industry leaders of automotive parts with an extensive background specializing in some of the best brands in the industry.
Our unique products and the highest quality customer service make us the industry favorites. Our facility is located in White Stone New York.